app/main.py
"""RoboDesk FastAPI application entry point."""

from dotenv import load_dotenv
load_dotenv()

from fastapi import FastAPI
from app.api.routes import router

app = FastAPI(
    title="RoboDesk — Axiom Robotics AI Support",
    description="RAG-powered customer service assistant for Axiom Robotics",
    version="1.0.0",
)

app.include_router(router)

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
