pyproject.toml
[project]
name = "robodesk"
version = "1.0.0"
description = "RoboDesk — AI-powered customer service for Axiom Robotics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-chroma>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
]
